fault tolerance and load balancing work with memoization/templates updated the job manager and version manager to make fault tolerance work with memoization/templates. if a shadow jobs is saved for a checkpoint, then explicit binding happens. also overwritten the read/write/union set methods for shadow job so that when saving the sets it reads from the pointer of template job, normaly the sets are empty for fast creation of shadow jobs.

+++ scheduler/checkpoint_entry.h
@@ -94,7 +94,7 @@ class CheckpointEntry {
     Index index_;
     JobEntryMap jobs_;
     checkpoint_id_t checkpoint_id_;
-    int pending_count_;
+    int64_t pending_count_;
 
     void IncreasePendingCounter();
